+++ source/telemetry/release-notes/release-notes.md
@@ -1,5 +1,9 @@
 # Telemetry Release Notes
 
+## Version 2.4.0
+
+- Added integration test for telemetry
+
 ## Version 2.1.2
 
 - Fix
@@ -10,7 +14,7 @@
 
 ## Version 2.1.0
 
-- Added shared telemtry library for logging
+- Added shared telemetry library for logging
 
 ## Version 1.9.5